+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.28.4
+
+### Fixed
+
+- Handle null rootUri's by not creating a Jedi project, fixes <https://github.com/pappasam/jedi-language-server/issues/95>
+- Tolerate invalid InitializationOptions by using default values on error
+
 ## 0.28.3
 
 ### Fixed